+++ lib/HLSL/CMakeLists.txt
@@ -10,9 +10,12 @@ add_llvm_library(LLVMHLSL
   DxilEliminateOutputDynamicIndexing.cpp
   DxilExpandTrigIntrinsics.cpp
   DxilGenerationPass.cpp
+  DxilLegalizeEvalOperations.cpp
   DxilLegalizeSampleOffsetPass.cpp
   DxilLinker.cpp
+  DxilPrecisePropagatePass.cpp
   DxilPreparePasses.cpp
+  DxilPromoteResourcePasses.cpp
   DxilPackSignatureElement.cpp
   DxilPatchShaderRecordBindings.cpp
   DxilPreserveAllOutputs.cpp
@@ -20,14 +23,17 @@ add_llvm_library(LLVMHLSL
   DxilSignatureValidation.cpp
   DxilTargetLowering.cpp
   DxilTargetTransformInfo.cpp
+  DxilTranslateRawBuffer.cpp
   DxilExportMap.cpp
   DxilValidation.cpp
   DxcOptimizer.cpp
+  HLDeadFunctionElimination.cpp
   HLExpandStoreIntrinsics.cpp
   HLMatrixBitcastLowerPass.cpp
   HLMatrixLowerPass.cpp
   HLMatrixSubscriptUseReplacer.cpp
   HLMatrixType.cpp
+  HLMetadataPasses.cpp
   HLModule.cpp
   HLOperations.cpp
   HLOperationLower.cpp
